feat(page): home
- Replacement method to stateless widgets.
