+++ lib/widgets/result_big_cards.dart
@@ -1,0 +1,144 @@
+import 'package:flutter/material.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:get/get.dart';
+import 'package:renting_house_training/core/constantes.dart';
+import 'package:renting_house_training/pages/details_page.dart';
+
+class ResultBigCards extends StatelessWidget {
+  const ResultBigCards({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: [
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            Text('Near from you', style: kheadlineMedium),
+            const Text('See more'),
+          ],
+        ),
+        SingleChildScrollView(
+          scrollDirection: Axis.horizontal,
+          child: Row(
+            children: List.generate(6, (int index) => buildingCard(index: index)),
+          ),
+        ),
+      ],
+    );
+  }
+
+  SizedBox buildingCard({required int index}) {
+    return SizedBox(
+      width: 272,
+      height: 320,
+      child: Card(
+        child: Stack(
+          children: [
+            SizedBox.expand(
+              child: Image.network(
+                'https://picsum.photos/seed/${{index + 1}}/200/300',
+                fit: BoxFit.cover,
+              ),
+            ),
+
+            /// Filter
+            SizedBox.expand(
+              child: InkWell(
+                onTap: () {
+                  Get.to<DetailsPage>(() => DetailsPage(index: index));
+                },
+                child: Container(
+                  decoration: BoxDecoration(
+                    gradient: LinearGradient(
+                      colors: [
+                        Colors.transparent,
+                        Colors.transparent,
+                        Colors.black.withValues(alpha: .5),
+                      ],
+                      begin: Alignment.topCenter,
+                      end: Alignment.bottomCenter,
+                    ),
+                  ),
+                ),
+              ),
+            ),
+            infoLocationCard(),
+            Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 16),
+              child: Align(
+                alignment: AlignmentDirectional.bottomStart,
+                child: Padding(
+                  padding: const EdgeInsets.all(8),
+                  child: Padding(
+                    padding: const EdgeInsets.all(8),
+                    child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Padding(
+                          padding: const EdgeInsets.symmetric(horizontal: 4),
+                          child: Text(
+                            'Dreamsville House',
+                            style: kTextButtonMedium,
+                          ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.symmetric(horizontal: 4),
+                          child: Text(
+                            'J.L Sultan Inskandar Muda',
+                            style: kBodyMedium.copyWith(color: kLightColor),
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  Widget infoLocationCard() {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 16),
+      child: Align(
+        alignment: AlignmentDirectional.topEnd,
+        child: Padding(
+          padding: const EdgeInsets.all(8),
+          child: Container(
+            decoration: BoxDecoration(
+              color: kShadowColor.withValues(alpha: .5),
+              borderRadius: BorderRadius.circular(40),
+            ),
+            child: Padding(
+              padding: const EdgeInsets.all(8),
+              child: Row(
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  const Padding(
+                    padding: EdgeInsets.symmetric(horizontal: 4),
+                    child: Icon(
+                      FontAwesomeIcons.locationDot,
+                      color: kLightColor,
+                    ),
+                  ),
+                  Padding(
+                    padding: const EdgeInsets.symmetric(horizontal: 4),
+                    child: Text(
+                      '1.8 km',
+                      style: kTextButtonMedium,
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
